refactor services to encapsulate database and caching, add migrations and logic
