+++ internal/services/cache/cache.go
@@ -1,4 +1,4 @@
-package services
+package cache
 
 import (
 	"context"
