+++ internal/services/database/database.go
@@ -1,4 +1,4 @@
-package services
+package database
 
 import (
 	"database/sql"
