+++ makefile
@@ -7,3 +7,9 @@ run: build
 
 watch:
 	air
+
+migrate-up:
+	migrate -path ./migrations -database "$(MIGRATION_DB_URL)" up
+
+migrate-down:
+	migrate -path ./migrations -database "$(MIGRATION_DB_URL)" down
